feat: eudirectlapse v3 vs v3.5 paired re-test — published loss does not reproduce (#186)
Tier 2 of the §15 re-test. Both model versions ran same-day in one environment
on folds identical to the published lapse benchmark, giving the project's first
paired version-vs-version test. v3.5 improves on v3 for log loss, Brier and
PR-AUC. The published v3 loss (AUC 0.6101 vs LR 0.6260) does not reproduce: v3
scores 0.6331 here and beats a one-hot LR on ranking; v3.5 beats it on all five
metrics. Cause of the August gap is unresolved (harness preprocessing vs
v3_default alias drift) and documented as such.

Also fixes the prediction read-back check, which scored float32-stored
predictions against float64 in-run scores; log_loss clips at dtype epsilon, so
hard 0/1 random-forest misses aborted both runs.

+++ scripts/eval/insurance_benchmark_v1/run_frontier_benchmark.py
@@ -819,7 +819,10 @@ def verify_predictions_readback(npz_path: Path, methods: list[str], metric, rows
         fold_vals = []
         for k in range(n_folds):
             y_true = data[f"y_true__fold{k}"]
-            y_pred = data[f"{m}__fold{k}"]
+            # Score in float64, as the in-run metric did: log_loss clips at the dtype's
+            # machine eps, so a hard 0/1 miss (RF pure leaves) costs ~16 nats from float32
+            # storage vs ~36 in-run — a 2.6e-3 mean shift on eudirectlapse from 3 misses.
+            y_pred = data[f"{m}__fold{k}"].astype(np.float64)
             if problem_type == "classification":
                 fold_vals.append(metric(y_true, _reconstruct_pp(m, y_pred)))
             else:
